Disable display of detailed exceptions
`@allow-rescue` permits rescuing exceptions by setting
`action_dispatch.show_exceptions` to true in the env hash. However,
this will render the detailed error page and not the public error page
you'd encounter in production.

Given that the feature enables testing of error pages, and seems to be
intended to mirror production behaviour as per the comment on
https://github.com/cucumber/cucumber-rails/blob/master/lib/generators/cucumber/install/templates/support/_rails_each_run.rb.erb#L6-L14,
this PR sets `action_dispatch.show_detailed_exceptions` to false, thus
enabling the public error page to be rendered.