Ensure selenium-wedriver 3 is not selected for tests
Since version 3 it requires geckodriver to be installed. This is not
available on Travis without requiring durther setup. Here we try to
restrict it in order to make the build pass.

Locally with geckodriver installed and selenium webdriver 3.4.0
everything is passing also. If this fails, the next step is to install
geckodriver on Travis.