Add support for Rails edge
We're in the process of getting onto Rails edge and this gem is one of
our dependencies that needs to be updated to allow running rails > 5.

I've opted for "rails-edge" over 6 here as can be left as is once rails
6 is released and people will be able to continue upgrading to the next
version with less friction. I can be swayed to do "rails-6" if you'd
prefer though :)