Runtime Dependency bump
Capybara had a lot of new matchers added / fixed in 2.12 (Primarily around frames)

This version is also still sufficiently old enough to support people who are running old ruby/rails combinations (Such as 2.3 ruby or 4.2 rails). But a big enough bump to suggest to people they really need to start upping their dependencies

Mime-Types is currently running on a restriction that supports software that was released before the advent of the first stable release of cucumber/cucumber-rails. It has been bumped to the next major version as a minimum (Which is still a large amount of years/commits behind. It will be bumped again down the line