https://github.com/cucumber/cucumber-rails/commit/a3d03e5df92f13a4bfc166a35c68a7427e9e365d#diff-9eec7a23bd3ec4020b3ff843a87c0e4a should have broken the build based on changes made in https://github.com/cucumber/cucumber-rails/commit/5df910110d44221ef5524abdd8df3b0f497e14da#diff-9eec7a23bd3ec4020b3ff843a87c0e4a
As the build was never broken, one can only assume the changes made at the time were based on an understanding Rails 5.1 did 'something' which in fact it didn't.

I've removed this logic but left the two sanitizer methods with a comment. These can be removed in cucumber-rails 2.1 if no further issues are detected. However it could be that this might raise an issue. However as this is only internal testing code, any issues should be negligible.

Obviously as this suite is reasonably old and used in a variety of places, there could be issues. In which case reverting this commit would suffice. However, this should not be necessary as the code should never work (Modifying frozen strings)

On-top of this, a minor refactor has been made to use better syntax for the auto-generated files (Later Ruby syntax)